Create user interface specification document
Asked 
"Please prepare the user interface specification document in English for the user management screen below (Markdown syntax preferred). It should include the requirements, details related to UI components, the behavior of the page when using these components, what to show to the user at the beginning, etc. This document will be used by Software developers who will develop this user interface. Please push the result in one of GitHub, GitLab, or BitBucket repositories (any of them is fine) and share the accessible link."